refactor(ES-013): gateway/api.ts をドメイン別ファイルに分割
2668行の api.ts を dispatcher + 9ドメインファイルに分割。
typecheck PASS / 655 tests PASS / madge 循環依存 0件。

- gateway/api/utils.ts — 共通ヘルパー (readBody, json, matchRoute 等)
- gateway/api/session-runner.ts — runWebSession / transcript / dispatch helpers
- gateway/api/sessions.ts — /api/sessions* ハンドラ
- gateway/api/cron.ts — /api/cron* ハンドラ
- gateway/api/org.ts — /api/org* ハンドラ
- gateway/api/skills.ts — /api/skills* ハンドラ
- gateway/api/connectors.ts — /api/connectors* ハンドラ
- gateway/api/stt.ts — /api/stt* ハンドラ
- gateway/api/misc.ts — config/logs/activity/onboarding/files/goals/budgets/costs ハンドラ
- gateway/api.ts — dispatcher のみ (39行)
